+++ lib/view/screens/event/event_screen.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutterAcmFall/view/screens/event/add_event_screen.dart';
 import 'package:flutterAcmFall/view/screens/event/event_setting_screen.dart';
+import 'package:flutterAcmFall/view/widget/event_list.dart';
 import 'package:flutterAcmFall/view/widget/event_utils.dart';
 import 'package:flutterAcmFall/model/objects/Event.dart';
 
@@ -41,7 +42,7 @@ class _EventScreen extends State<EventScreen> {
     });
   }
 
-  void _handleChosenCard(Event event) {
+  void _handleClickCard(Event event) {
     setState(() {
       _openSettingScreen = true;
       _cardEvent = event;
@@ -91,17 +92,11 @@ class _EventScreen extends State<EventScreen> {
               mainAxisAlignment: MainAxisAlignment.spaceBetween,
               children: <Widget>[
             Container(
-              height: MediaQuery.of(context).size.height * 0.5,
-              child: SingleChildScrollView(
-                  child: Column(
-                      children: _events
-                          .map((event) => EventCard(
-                                event: event,
-                                onClickCard: _handleChosenCard,
-                                onDeleteCard: _handleDeleteCard,
-                              ))
-                          .toList())),
-            ),
+                child: EventList(
+              events: _events,
+              handleClickEvent: _handleClickCard,
+              handleDeleteEvent: _handleDeleteCard,
+            )),
             Padding(
                 padding: EdgeInsets.symmetric(vertical: 15, horizontal: 10),
                 child: Container(
@@ -120,11 +115,13 @@ class _EventScreen extends State<EventScreen> {
                                       fontSize: 25,
                                       fontWeight: FontWeight.bold,
                                       color: Colors.white)),
-                              Text("${_events.length} events",
-                                  style: TextStyle(
-                                      fontSize: 18,
-                                      color:
-                                          Color.fromRGBO(244, 244, 244, 0.5)))
+                              _events.length > 0
+                                  ? Text("${_events.length} events",
+                                      style: TextStyle(
+                                          fontSize: 18,
+                                          color: Color.fromRGBO(
+                                              244, 244, 244, 0.5)))
+                                  : Container()
                             ]))))
           ])),
       floatingActionButton: FloatingActionButton(
@@ -139,100 +136,5 @@ class _EventScreen extends State<EventScreen> {
       _openAddEventScreen ? addEventScreen : Container(),
       settingEventScreen
     ]);
-  }
-}
-
-class EventCard extends StatefulWidget {
-  EventCard({Key key, this.event, this.onClickCard, this.onDeleteCard})
-      : super(key: key);
-
-  final Event event;
-  final Function onClickCard;
-  final Function onDeleteCard;
-
-  _EventCard createState() => _EventCard();
-}
-
-class _EventCard extends State<EventCard> {
-  @override
-  Widget build(BuildContext context) {
-    String title = (widget.event.title.length >= 25)
-        ? '${widget.event.title.substring(0, 25)}...'
-        : widget.event.title;
-
-    FontWeight titleWeight =
-        widget.event.isDone ? FontWeight.bold : FontWeight.normal;
-
-    Widget dateText = widget.event.date != null
-        ? Padding(
-            padding: EdgeInsets.only(top: 3.0, bottom: 3.0),
-            child: EventDateFormat(date: widget.event.date))
-        : Container();
-
-    Widget timeText = widget.event.time != null
-        ? Padding(
-            padding: EdgeInsets.only(top: 3.0, bottom: 3.0, left: 10.0),
-            child: EventTimeFormat(time: widget.event.time))
-        : Container();
-
-    return // debug
-        InkWell(
-            onTap: () {
-              print(widget.event);
-              widget.onClickCard(widget.event);
-            },
-            child: Container(
-                height: 70,
-                decoration: BoxDecoration(
-                    border: Border(
-                        bottom: BorderSide(
-                            color: Color.fromRGBO(235, 239, 245, 1.0),
-                            width: 2))),
-                child: InkWell(
-                    onTap: () {
-                      print(widget.event);
-                      widget.onClickCard(widget.event);
-                    },
-                    child: Padding(
-                        padding: EdgeInsets.symmetric(horizontal: 18.0),
-                        child: Row(children: <Widget>[
-                          RadioButton(
-                            onTap: () {
-                              setState(() {
-                                widget.event.isDone = !widget.event.isDone;
-                              });
-                            },
-                            isActive: widget.event.isDone,
-                          ),
-                          Expanded(
-                              child: Padding(
-                                  padding: EdgeInsets.symmetric(horizontal: 10),
-                                  child: Column(
-                                      mainAxisAlignment:
-                                          MainAxisAlignment.center,
-                                      crossAxisAlignment:
-                                          CrossAxisAlignment.start,
-                                      children: <Widget>[
-                                        Padding(
-                                            padding: EdgeInsets.symmetric(
-                                                vertical: 3.0),
-                                            child: Text(title,
-                                                style: TextStyle(
-                                                    fontSize: 20,
-                                                    fontWeight: titleWeight,
-                                                    color: Color.fromRGBO(
-                                                        37, 42, 49, 1.0)))),
-                                        Row(children: <Widget>[
-                                          dateText,
-                                          timeText
-                                        ]),
-                                      ]))),
-                          DeleteButton(
-                            onTap: () {
-                              widget.onDeleteCard(widget.event);
-                            },
-                            isActive: true,
-                          ),
-                        ])))));
   }
 }
